requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
python-multipart==0.0.9
sentence-transformers==2.6.1
faiss-cpu==1.7.4
numpy==1.26.4
pdfplumber==0.11.0
python-docx==1.1.0
pillow==10.3.0
pydantic==2.6.4
pytest==8.1.1
httpx==0.27.0
